# backend/src/main/java/ru/itmo/accontrol/dto/CreateUserRequest.java
package ru.itmo.accontrol.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CreateUserRequest {
    
    @NotBlank(message = "OAuth token is required")
    private String oauthToken;
}